Refactor onboarding types and completion logic
