+++ src/types/index.ts
@@ -318,17 +318,3 @@ export interface DailyStatistics {
   isRealTime: boolean;
   lastUpdated: string;
 }
-
-export const ONBOARDING_STEP = {
-  INFO: 'info',
-  TABLES: 'tables',
-  MENU: 'menu',
-  COMPLETE: 'complete',
-} as const;
-
-export type OnboardingStep = typeof ONBOARDING_STEP[keyof typeof ONBOARDING_STEP];
-
-export interface OnboardingStepDefinition {
-  step: OnboardingStep;
-  label: string;
-}
